Consolidate ingress_per_unit library to one file
* Merge the provides.py and requires.py into one file.
* Move the unit-testing of the library to the /tests folder
  of the charm.
* Remove @cachedproperty, as it caused stale data to be
  provided to the requirer.
